Fixing functionality after working through mern quickstart example
