+++ apps/mern-quickstart/.dockerignore
@@ -1,4 +1,4 @@
-// .dockerignore
+# .dockerignore
 
 node_modules
 npm-debug.log
