+++ apps/mern-quickstart/app.js
@@ -5,19 +5,26 @@ const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
 
 // initializing routes
-const users = require('./routes/api/users');
+const users = require('./routes/user');
 
 const app = express();
+
+// Body parser middleware
+app.use(bodyParser.urlencoded({ extended: false }));
+app.use(bodyParser.json());
 
 // DB config
 const db = require('./config/keys').mongoURI;
 
+const mongooseOptions = {
+  useNewUrlParser: true,
+  useUnifiedTopology: true,
+  useCreateIndex: true
+}
+
 // Connect to MongoDB
 mongoose
-  .connect(
-    db,
-    { useNewUrlparser: true }
-  )
+  .connect(db, mongooseOptions)
   .then(() => console.log('MongoDB Connected'))
   .catch((err) => console.error(err));
 
@@ -25,10 +32,6 @@ app.get('/', (req, res) => res.send('Hello World!'));
 
 // use Routes
 app.use('/api/users', users);
-
-// Body parser middleware
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(bodyParser.json());
 
 const port = process.env.PORT  || 8082;
