+++ apps/mern-quickstart/docker-compose.yml
@@ -4,6 +4,8 @@ services:
     container_name: quickstart-server
     restart: always
     build: .
+    volumes:
+      - .:/usr/src/app
     ports:
       - '80:8082'
     links:
