+++ apps/mern-quickstart/models/User.js
@@ -11,11 +11,12 @@ const UserSchema = new Schema({
   },
   email: {
     type: String,
-    required: true
+    required: true,
+    unique:true
   },
   password: {
-    tpe: String,
-    required: true
+    type: String,
+    required: true,
   },
   date: {
     type: Date,
@@ -23,4 +24,6 @@ const UserSchema = new Schema({
   }
 });
 
-module.exports = User = mongoose.model('users', UserSchema);
+const User = mongoose.model('users', UserSchema);
+
+module.exports = User;
